requirements.txt
crewai>=0.41.0
python-dotenv>=1.0.0
flask>=2.3.3
pyyaml>=6.0.1
tiktoken>=0.5.2
requests>=2.31.0
google-generativeai==0.7.2